refactor: ViewModel 팩토리 통합 및 상품 상태 구조 정리
