+++ app/src/main/java/woowacourse/shopping/ui/productlist/ProductListViewModel.kt
@@ -1,5 +1,6 @@
 package woowacourse.shopping.ui.productlist
 
+import android.os.SystemClock
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import kotlinx.coroutines.flow.MutableSharedFlow
@@ -9,19 +10,29 @@ import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asSharedFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
+import kotlinx.coroutines.sync.Mutex
+import kotlinx.coroutines.sync.withLock
 import woowacourse.shopping.data.local.datastore.VisitStore
+import woowacourse.shopping.data.mapper.toDomainProducts
+import woowacourse.shopping.data.mapper.toShoppingItem
+import woowacourse.shopping.data.remote.retrofit.awaitBody
+import woowacourse.shopping.data.remote.retrofit.repository.ProductRetrofitRepository
 import woowacourse.shopping.domain.model.ShoppingItem
 import woowacourse.shopping.domain.repository.ShoppingItemRepository
 
 class ProductListViewModel(
     private val shoppingItemRepository: ShoppingItemRepository,
     private val visitStore: VisitStore,
+    private val productRetrofitRepository: ProductRetrofitRepository,
 ) : ViewModel() {
     private val _uiState = MutableStateFlow(ProductListUiState())
     val uiState: StateFlow<ProductListUiState> = _uiState.asStateFlow()
     private val _event = MutableSharedFlow<ProductListEvent>(extraBufferCapacity = 1)
     val event: SharedFlow<ProductListEvent> = _event.asSharedFlow()
 
+    private val productRequestMutex = Mutex()
+    private var hasLoadedProductsOnce: Boolean = false
+    private var lastProductsLoadedElapsedMs: Long = 0L
     private val productPageStateHolder = ProductPageStateHolder(shoppingItems = emptyList())
     private var allShoppingItems: List<ShoppingItem> = shoppingItemRepository.shoppingItems.value
     private var recentViewedProductIds: List<Long> = visitStore.recentVisitedProductIds.value
@@ -30,6 +41,52 @@ class ProductListViewModel(
         productPageStateHolder.updateItems(allShoppingItems)
         publishUiState()
         observeSources()
+    }
+
+    fun requestProduct(
+        page: Int = DEFAULT_PAGE,
+        size: Int = DEFAULT_SIZE,
+        category: String? = null,
+        force: Boolean = false,
+    ) {
+        if (shouldSkipProductRequest(force = force)) return
+
+        viewModelScope.launch {
+            productRequestMutex.withLock {
+                if (shouldSkipProductRequest(force = force)) return@withLock
+
+                publishUiState(
+                    isLoading = true,
+                    errorMessage = null,
+                )
+
+                runCatching {
+                    productRetrofitRepository
+                        .requestProduct(
+                            page = page,
+                            size = size,
+                            category = category,
+                        ).awaitBody(errorPrefix = "상품 조회 실패")
+                        .toDomainProducts()
+                }.onSuccess { loadedProducts ->
+                    val loadedShoppingItems = loadedProducts.map { product -> product.toShoppingItem() }
+                    if (loadedProducts.isNotEmpty()) {
+                        shoppingItemRepository.replaceProducts(loadedProducts)
+                    }
+                    markProductsLoaded()
+                    publishUiState(
+                        isLoading = false,
+                        errorMessage = null,
+                        loadedProducts = loadedShoppingItems,
+                    )
+                }.onFailure {
+                    publishUiState(
+                        isLoading = false,
+                        errorMessage = "상품 목록을 불러오지 못했습니다.",
+                    )
+                }
+            }
+        }
     }
 
     fun loadNextPage() {
@@ -69,13 +126,29 @@ class ProductListViewModel(
         }
     }
 
-    private fun publishUiState() {
-        _uiState.value = createUiState()
+    private fun publishUiState(
+        isLoading: Boolean = _uiState.value.isLoading,
+        errorMessage: String? = _uiState.value.errorMessage,
+        loadedProducts: List<ShoppingItem> = _uiState.value.loadedProducts,
+    ) {
+        _uiState.value =
+            createUiState(
+                isLoading = isLoading,
+                errorMessage = errorMessage,
+                loadedProducts = loadedProducts,
+            )
     }
 
-    private fun createUiState(): ProductListUiState {
+    private fun createUiState(
+        isLoading: Boolean,
+        errorMessage: String?,
+        loadedProducts: List<ShoppingItem>,
+    ): ProductListUiState {
         val shoppingItemByProductId = allShoppingItems.associateBy { shoppingItem -> shoppingItem.getProductId() }
         return ProductListUiState(
+            isLoading = isLoading,
+            errorMessage = errorMessage,
+            loadedProducts = loadedProducts,
             shoppingItems = productPageStateHolder.getItems(),
             recentViewedShoppingItems = recentViewedProductIds.mapNotNull { productId -> shoppingItemByProductId[productId] },
             shoppingCartTotalCount = allShoppingItems.sumOf { shoppingItem -> shoppingItem.getQuantity() },
@@ -83,7 +156,24 @@ class ProductListViewModel(
         )
     }
 
+    private fun shouldSkipProductRequest(force: Boolean): Boolean {
+        if (force) return false
+        if (!hasLoadedProductsOnce) return false
+        return isProductsCacheFresh()
+    }
+
+    private fun isProductsCacheFresh(): Boolean =
+        SystemClock.elapsedRealtime() - lastProductsLoadedElapsedMs < PRODUCTS_CACHE_DURATION_MS
+
+    private fun markProductsLoaded() {
+        hasLoadedProductsOnce = true
+        lastProductsLoadedElapsedMs = SystemClock.elapsedRealtime()
+    }
+
     data class ProductListUiState(
+        val isLoading: Boolean = false,
+        val errorMessage: String? = null,
+        val loadedProducts: List<ShoppingItem> = emptyList(),
         val shoppingItems: List<ShoppingItem> = emptyList(),
         val recentViewedShoppingItems: List<ShoppingItem> = emptyList(),
         val shoppingCartTotalCount: Int = 0,
@@ -97,5 +187,11 @@ class ProductListViewModel(
         ) : ProductListEvent
 
         data object NavigateToShoppingCart : ProductListEvent
+    }
+
+    private companion object {
+        private const val DEFAULT_PAGE = 0
+        private const val DEFAULT_SIZE = 20
+        private const val PRODUCTS_CACHE_DURATION_MS = 30_000L
     }
 }
